# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "CategoryOverrideRequest": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/StorageCategory"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "``None`` clears the override.",
            "title": "Path"
          }
        },
        "required": [
          "category"
        ],
        "title": "CategoryOverrideRequest",
        "type": "object"
      },
      "CategoryUsage": {
        "properties": {
          "bytes_used": {
            "title": "Bytes Used",
            "type": "integer"
          },
          "category": {
            "$ref": "#/components/schemas/StorageCategory"
          },
          "exists": {
            "title": "Exists",
            "type": "boolean"
          },
          "file_count": {
            "title": "File Count",
            "type": "integer"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "protected": {
            "description": "Warn before deleting (checkpoints, training data, memory).",
            "title": "Protected",
            "type": "boolean"
          }
        },
        "required": [
          "category",
          "path",
          "bytes_used",
          "file_count",
          "exists",
          "protected"
        ],
        "title": "CategoryUsage",
        "type": "object"
      },
      "CleanupCandidate": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/StorageCategory"
          },
          "kind": {
            "$ref": "#/components/schemas/CleanupKind"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "protected": {
            "description": "Under a category that warns before deletion.",
            "title": "Protected",
            "type": "boolean"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "path",
          "category",
          "kind",
          "size_bytes",
          "protected",
          "reason"
        ],
        "title": "CleanupCandidate",
        "type": "object"
      },
      "CleanupKind": {
        "enum": [
          "partial_download",
          "orphaned_model"
        ],
        "title": "CleanupKind",
        "type": "string"
      },
      "CleanupPlan": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CleanupCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "protected_bytes": {
            "title": "Protected Bytes",
            "type": "integer"
          },
          "reclaimable_bytes": {
            "title": "Reclaimable Bytes",
            "type": "integer"
          }
        },
        "required": [
          "candidates",
          "reclaimable_bytes",
          "protected_bytes"
        ],
        "title": "CleanupPlan",
        "type": "object"
      },
      "CleanupRequest": {
        "properties": {
          "acknowledge_protected": {
            "default": false,
            "description": "Must be true to delete anything under a protected category.",
            "title": "Acknowledge Protected",
            "type": "boolean"
          },
          "paths": {
            "items": {
              "type": "string"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Paths",
            "type": "array"
          }
        },
        "required": [
          "paths"
        ],
        "title": "CleanupRequest",
        "type": "object"
      },
      "CleanupResult": {
        "properties": {
          "deleted": {
            "items": {
              "type": "string"
            },
            "title": "Deleted",
            "type": "array"
          },
          "freed_bytes": {
            "title": "Freed Bytes",
            "type": "integer"
          },
          "refused": {
            "description": "Paths that were not deleted and why.",
            "items": {
              "type": "string"
            },
            "title": "Refused",
            "type": "array"
          }
        },
        "required": [
          "deleted",
          "freed_bytes",
          "refused"
        ],
        "title": "CleanupResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StorageCategory": {
        "enum": [
          "core",
          "models",
          "skills",
          "training",
          "checkpoints",
          "memory",
          "knowledge",
          "projects",
          "generated"
        ],
        "title": "StorageCategory",
        "type": "string"
      },
      "StorageLocationCheck": {
        "description": "Result of validating a candidate storage root before committing to it.",
        "properties": {
          "free_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Bytes"
          },
          "is_existing_myai_storage": {
            "default": false,
            "title": "Is Existing Myai Storage",
            "type": "boolean"
          },
          "is_removable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Removable"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "problems": {
            "items": {
              "type": "string"
            },
            "title": "Problems",
            "type": "array"
          },
          "total_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Bytes"
          }
        },
        "required": [
          "path",
          "ok",
          "problems",
          "is_existing_myai_storage",
          "free_bytes",
          "total_bytes",
          "is_removable"
        ],
        "title": "StorageLocationCheck",
        "type": "object"
      },
      "StorageOverview": {
        "properties": {
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryUsage"
            },
            "title": "Categories",
            "type": "array"
          },
          "category_overrides": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Category Overrides",
            "type": "object"
          },
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "configured_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configured At"
          },
          "root_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Path"
          },
          "total_bytes_used": {
            "title": "Total Bytes Used",
            "type": "integer"
          },
          "volume_free_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume Free Bytes"
          },
          "volume_total_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume Total Bytes"
          }
        },
        "required": [
          "configured",
          "root_path",
          "categories",
          "total_bytes_used",
          "volume_total_bytes",
          "volume_free_bytes",
          "configured_at",
          "category_overrides"
        ],
        "title": "StorageOverview",
        "type": "object"
      },
      "StorageSetupRequest": {
        "properties": {
          "root_path": {
            "minLength": 1,
            "title": "Root Path",
            "type": "string"
          }
        },
        "required": [
          "root_path"
        ],
        "title": "StorageSetupRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "MyAI Academy local service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/storage/check": {
      "get": {
        "operationId": "check_location_api_storage_check_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageLocationCheck"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Location",
        "tags": [
          "storage"
        ]
      }
    },
    "/api/storage/cleanup": {
      "get": {
        "description": "What could be deleted safely, with protected items flagged (spec \u00a763).",
        "operationId": "cleanup_plan_api_storage_cleanup_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupPlan"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Cleanup Plan",
        "tags": [
          "storage"
        ]
      },
      "post": {
        "operationId": "cleanup_api_storage_cleanup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CleanupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cleanup",
        "tags": [
          "storage"
        ]
      }
    },
...
    "/api/storage/overrides": {
      "put": {
        "operationId": "set_override_api_storage_overrides_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryOverrideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Override",
        "tags": [
          "storage"
        ]
      }
    },
    "/api/storage/root": {
      "put": {
        "operationId": "configure_root_api_storage_root_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorageSetupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Configure Root",
        "tags": [
          "storage"
        ]
      }
    }
  }
}